2025_06_13_ver_9

+++ s21_atoi.h
@@ -6,6 +6,7 @@
 char *s21_atoi(param *param);
 char *s21_atof(param *param);
 char *s21_atoi_new(param *param);
-int s21_atoi_int(char *str, char *digit, long long int x, param *param);
+//int s21_atoi_int(char *str, char *digit, long long int x, param *param);
+int s21_atoi_int(char *str, char *digit, long double x, param *param);
 
 #endif
